requirements.txt
discord.py>=2.0
Flask
python-dotenv
Pillow
requests
